requirements.txt
absl-py
aiohttp
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs
blinker==1.4
brotlipy==0.7.0
cachetools
certifi==2021.5.30
cffi
chardet
charset-normalizer
click
coverage
cryptography
cycler==0.10.0
Cython
Flask==2.0.1
flatbuffers
gast
gevent==21.8.0
google-auth
google-auth-oauthlib
google-pasta
greenlet==1.1.2
grpcio
h5py
idna
importlib-metadata
itsdangerous==2.0.1
Jinja2==3.0.1
Keras
Keras-Preprocessing
kiwisolver==1.3.2
Markdown
MarkupSafe==2.0.1
matplotlib==3.4.3
mkl-fft==1.3.0
mkl-random
mkl-service==2.4.0
multidict
nltk==3.6.2
numpy
oauthlib
opt-einsum
Pillow==8.3.2
protobuf==3.17.2
pyasn1
pyasn1-modules==0.2.8
pycparser
PyJWT
pyOpenSSL
pyparsing==2.4.7
PySocks
python-dateutil==2.8.2
PyYAML==5.4.1
requests
requests-oauthlib==1.3.0
rsa
scipy
six
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor==1.1.0
typing-extensions
urllib3
Werkzeug
wrapt==1.12.1
yarl
zipp
zope.event==4.5.0
zope.interface==5.4.0
